+++ priv/repo/migrations/20190118134400_remove_template_relation.exs
@@ -1,0 +1,7 @@
+defmodule TdDf.Repo.Migrations.RemoveTemplateRelation do
+  use Ecto.Migration
+
+  def change do
+    drop table(:template_relations)
+  end
+end